New files
Added a copy of Starter Book. The original one is not showing the correct line chart so I went back and tried to recreate it but that did not change the results so I am starting a new one. Once this is working I can use it for the basis of the book for the challenge.